refactor: Introduce file list
